pyproject.toml
[tool.poetry]
name = "shoping-cart"
version = "0.1.0"
description = ""
authors = ["Lev Alekseev <futuristjp@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
black = "^23.12.1"
django-filter = "^23.5"
psycopg2 = "^2.9.9"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
